+++ packages/ilmomasiina-backend/src/models/migrations/0011-add-settings.ts
@@ -1,0 +1,37 @@
+import { DataTypes } from "sequelize";
+
+import { defineMigration } from "./util";
+
+export default defineMigration({
+  name: "0011-add-settings",
+  async up({ context: { sequelize, transaction } }) {
+    const query = sequelize.getQueryInterface();
+    await query.createTable(
+      "setting",
+      {
+        key: {
+          type: DataTypes.STRING(64),
+          allowNull: false,
+          primaryKey: true,
+        },
+        value: {
+          type: DataTypes.JSONB,
+          allowNull: false,
+        },
+        createdAt: {
+          type: DataTypes.DATE,
+          allowNull: false,
+        },
+        updatedAt: {
+          type: DataTypes.DATE,
+          allowNull: false,
+        },
+      },
+      { transaction },
+    );
+  },
+  async down({ context: { sequelize, transaction } }) {
+    const query = sequelize.getQueryInterface();
+    await query.dropTable("setting", { transaction });
+  },
+});